+++ apps/pie-microsite/src/_11ty/shortcodes/tokensTable/index.js
@@ -3,10 +3,11 @@ const normalizedPieDesignTokens = require('../../../_data/normalizeTokens');
 const pieTokenCategories = require('../../../tokenCategories.json');
 const { stringHelpers, objectHelpers, numberHelpers } = require('../../../utilities/helpers');
 const tokenTypes = require('../../../_data/tokenTypes');
-const { buildColorName, buildColorExample } = require('./tokenTypes/colour');
+const { buildColorName, buildColorExample, buildColorDescription } = require('./tokenTypes/colour');
 const { buildSpacingExample } = require('./tokenTypes/spacing');
 const { buildFontExample } = require('./tokenTypes/font');
 const { buildRadiusExample } = require('./tokenTypes/radius');
+const { deindentHTML } = require('../shortcode-utilities');
 
 const {
     getParentCategoriesForTokenType,
@@ -65,6 +66,47 @@ const buildTokenExampleElement = (token, tokenType, tokenMetadata) => {
     return tokenExampleElementHandler[tokenType](token, tokenMetadata);
 };
 
+/**
+ * Builds the global token used element displayed primarily by alias tokens
+ * @param {string} globalToken the global token referenced by this alias token
+ * @returns a <span> HTML string containing the global token used
+ */
+const buildGlobalTokenUsedElement = globalToken => {
+    const globalTokenUsedElement = `
+    <span class="c-tokensTable-tokenDescription">
+      <span class="u-font-bold u-showAboveWide">Global token used:</span> 
+      <span class="c-tokensTable-token c-tokensTable-token--light">${globalToken}</span>
+    </span>`;
+
+    return deindentHTML(globalTokenUsedElement);
+};
+
+/**
+ * Builds the overall token description element for each type of token. The description content differs based on the type of token.
+ * @param {string} token the token value i.e. #000, #ffffff, #000|0.85 or #000000|0.85
+ * @param {*} tokenType the type of token i.e. color, spacing, radius
+ * @param {*} tokenMetadata the metadata for the token. data such as descriptions
+ * @returns {string} - the description HTML string
+ */
+const buildTokenDescriptionElement = (token, tokenType, tokenMetadata) => {
+    const tokenTypeBuilder = {
+        [tokenTypes.COLOR]: buildColorDescription,
+        default: () => (tokenMetadata.description
+            ? `<span class="c-tokensTable-tokenDescription">${tokenMetadata.description}</span>`
+            : '')
+    };
+
+    let description = tokenType in tokenTypeBuilder
+        ? tokenTypeBuilder[tokenType](token, tokenMetadata)
+        : tokenTypeBuilder.default(tokenMetadata);
+
+    if (tokenMetadata.globalToken) {
+        description += buildGlobalTokenUsedElement(tokenMetadata.globalToken);
+    }
+
+    return description;
+};
+
 
 /**
  * Builds a token pill element to display the SCSS token name in the token list item.
@@ -96,17 +138,17 @@ const buildTokenListElements = ({
     // TODO - description is just an example of how we might use the metadata
     // We would likely wanted to move them into a colour specific handler similar to how we build
     // the colour token example. Please consider them placeholder for now.
-    const tokenDescription = tokenMetadata.description
-        ? `<span class="c-tokensTable-tokenDescription">${tokenMetadata.description}</span>`
-        : '';
+    const tokenDescription = buildTokenDescriptionElement(token, tokenType, tokenMetadata);
 
-    return `<li class="c-tokensTable-row c-tokensTable-item">
+    return deindentHTML(`
+    <li class="c-tokensTable-row c-tokensTable-item">
       ${tokenExampleElement}
       <div class="c-tokensTable-content">
-        <span class="c-tokensTable-displayName">${tokenDisplayName}</span>${tokenDescription}
+        <span class="c-tokensTable-displayName">${tokenDisplayName}</span>
+        ${tokenDescription}
       </div>
       ${tokenPill}
-    </li>`;
+    </li>`);
 };
 
 /**
@@ -188,7 +230,7 @@ const buildCategorisedLists = ({
 
     // for each category, create an h2 and a list of token elements to render
     const categoryTokenLists = Object.keys(categories).map(category => {
-        const heading = `<h2>${categories[category].displayName}</h2>`;
+        const heading = `<h2 class="c-tokensTable-title">${categories[category].displayName}</h2>`;
         const tokensList = buildTokensListForCategory(tokens, path, category, tokenType);
 
         // returns a 'chunk' of the tokens table page (a heading, the column headers, list of tokens and an option HR element)
@@ -225,7 +267,7 @@ const buildCategoryListsWithParents = ({
 
         // create heading and description for parent category
         const { displayName, description } = parentCategories[parentCategoryKey];
-        const heading = `<h2 class="c-tokensTable-sectionHeading">${displayName}</h2>`;
+        const heading = `<h2 class="c-tokensTable-title c-tokensTable-sectionHeading">${displayName}</h2>`;
         const descriptionMarkup = `<p class="c-tokensTable-sectionDescription">${description}</p>`;
 
         // return heading, description and lists