+++ apps/pie-microsite/CHANGELOG.md
@@ -3,6 +3,18 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+v1.47.0
+------------------------------
+*January 26, 2023*
+
+### Added
+- Alias token descriptions for colour page
+
+### Updated
+- `@justeat/pie-design-tokens v3.7.0`
+
+
 v1.46.1
 ------------------------------
 *January 26, 2023*